Modify run script to take advantage of JeOS
Modify the virt assistant function to:

 1) Download the JeOS image
 2) Setup the environment non interactively

So it all will be done without user intervention.
The first time the JeOS will have to be built,
a couple of minutes to download the 120MB compressed
image + 1 minute to uncompress the image will be
required, after that it'll keep using the uncompressed
image, unless you want to replace it by the pristine
file you have on your hard drive, then you have to call
run with --restore-image.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>